authorization/requirements.txt
PyJWT==2.9.0
python-dotenv==1.0.1
redis==5.1.1
starlette==0.41.0
strawberry==3.0
uvicorn==0.32.0